mcp_server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0
websockets>=11.0.3
httpx>=0.24.1
groq>=0.4.0
google-generativeai>=0.3.0
langchain>=0.0.267
langchain-groq>=0.0.1
langchain-google-genai>=0.0.1
